requirements.txt
Flask==2.2.5
Flask_Cors==4.0.1
matplotlib==3.3.2
numpy==1.16.0
numpy==1.21.6
opencv_contrib_python==4.9.0.80
opencv_python_headless==4.8.1.78
Pillow==9.4.0
Pillow==9.5.0
Pillow==10.3.0
torch==1.13.1
torchvision==0.14.1
